+++ CashRegister.Infrastructure/Migrations/CashRegisterDBContextModelSnapshot.cs
@@ -34,7 +34,7 @@ namespace CashRegister.Infrastructure.Migrations
                         .IsRequired()
                         .HasColumnType("text");
 
-                    b.Property<decimal>("TotalPrice")
+                    b.Property<int>("TotalPrice")
                         .HasColumnType("numeric");
 
                     b.HasKey("BillNumber");
@@ -54,7 +54,7 @@ namespace CashRegister.Infrastructure.Migrations
                         .IsRequired()
                         .HasColumnType("text");
 
-                    b.Property<decimal>("Price")
+                    b.Property<int>("Price")
                         .HasColumnType("numeric");
 
                     b.HasKey("Id");
@@ -73,7 +73,7 @@ namespace CashRegister.Infrastructure.Migrations
                     b.Property<int>("ProductQuantity")
                         .HasColumnType("integer");
 
-                    b.Property<decimal>("ProductsPrice")
+                    b.Property<int>("ProductsPrice")
                         .HasColumnType("numeric");
 
                     b.HasKey("BillNumber", "ProductId");
